auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg[binary]==3.1.18
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.8
cryptography==41.0.7
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
aiosmtplib==3.0.1
